Initial commit
This is the non functioning version. I initially started playing with this idea over a year ago. Last modified date was 2018-10-18 when I started fiddling wit hthis yesterday, while it was still using the dbus npm package. Unfortunately that package is a broken mess and it didn't work, for whatever reason. I started fiddling with it again yesterday and managed to get a working version by changing the out.type to be an object instead of an array. Then today, while trying to make notifications work without making the program unable to receive any more match requests (yes, the package is that broken) I decided to migrate to dbus-native. Now both KRunner integration and notifications work, except KRunner doesn't actually display any of the results. Unfortunately I have no idea why that is, so it's time to migrate back to dbus to check if the replies I get over DBus aren't different in some significant way. Oh well.

Some notes on getting this to run: the desktop file has to be visible to KRunner for it to send any messages to us over DBus. This can be achieved by linking the desktop file in `~/.local/share/kservices5/`. I should probably put this in a readme instead...